\name{fma-package}
\alias{fma-package}
\alias{fma}
\docType{package}
\title{
Data from the M-competitions
}
\description{The 1001 time series from the M-competition (Makridakis et al. 1982) 
  and the 3003 time series from the IJF-M3 competition (Makridakis and Hibon, 2000).
}

\author{Rob J Hyndman. \email{Rob.Hyndman@buseco.monash.edu.au}}

\source{
  \url{http://forecasters.org/time-series.html}.

  Detailed results from M3 competition at
  \url{http://forecastingprinciples.com/index.php?option=com_content&task=view&id=209}.
}

\references{
  Makridakis, S., A. Andersen, R. Carbone, R. Fildes, M. Hibon, R.
Lewandowski, J. Newton, E. Parzen, and R. Winkler (1982) The accuracy of
extrapolation (time series) methods: results of a forecasting competition.
\emph{Journal of Forecasting}, \bold{1}, 111--153.

  Makridakis and Hibon (2000) The M3-competition: results,
  conclusions and implications. \emph{International Journal of Forecasting}, \bold{16}, 451-476.
}


\keyword{ package }


