% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mcomp-package.R
\docType{package}
\name{Mcomp-package}
\alias{Mcomp-package}
\alias{Mcomp}
\title{Data from the M-competitions}
\source{
\url{http://forecasters.org/resources/time-series-data/m3-competition/}.

Detailed results from M3 competition at
\url{http://www.insead.edu/facultyresearch/research/doc.cfm?did=1094}.
}
\description{
The 1001 time series from the M-competition (Makridakis et al. 1982), and
the 3003 time series and forecasts from the IJF-M3 competition (Makridakis
and Hibon, 2000).
}
\references{
Makridakis, S., A. Andersen, R. Carbone, R. Fildes, M. Hibon, R.
Lewandowski, J. Newton, E. Parzen, and R. Winkler (1982) The accuracy of
extrapolation (time series) methods: results of a forecasting competition.
\emph{Journal of Forecasting}, \bold{1}, 111--153.

Makridakis and Hibon (2000) The M3-competition: results, conclusions and
implications. \emph{International Journal of Forecasting}, \bold{16},
451-476.
}
\author{
Rob J Hyndman. \email{Rob.Hyndman@monash.edu}, with assistance from
Muhammad Akram and Christoph Bergmeir.
}
\keyword{package}
