% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.Mdata}
\alias{plot.Mdata}
\alias{autoplot.Mdata}
\title{Plotting M Competition data}
\usage{
\method{plot}{Mdata}(x, xlim = c(tsp(x$x)[1], tsp(x$xx)[2]),
  ylim = range(x$x, x$xx), main = x$sn, xlab, ylab = "", ...)

\method{autoplot}{Mdata}(object, ...)
}
\arguments{
\item{x, object}{A series of M-competition data}

\item{xlim}{Limits on x-axis}

\item{ylim}{Limits on y-axis}

\item{main}{Main title}

\item{xlab}{Label on x-axis}

\item{ylab}{Label on y-axis}

\item{...}{Other plotting arguments passed to \code{plot}. Ignored for
\code{autoplot}.}
}
\value{
\code{autoplot.Mdata} returns a ggplot2 object, while
\code{plot.Mdata} returns nothing. Both functions produce a time series plot
of the selected series.
}
\description{
Functions to plot a time series from the M competition data sets, showing
both the training and test sections of the series.
}
\examples{

library(ggplot2)
plot(M1[[1]])
autoplot(M1$YAF3)
autoplot(M3[["N0647"]])
}
\seealso{
\code{\link{M1}}, \code{\link{M3}}
}
\author{
Rob Hyndman
}
\keyword{hplot}
